+++ lib/simply_stored/couch/has_many.rb
@@ -2,6 +2,7 @@ module SimplyStored
   module Couch
     module HasMany
       def has_many(name, options = {})
+        check_existing_properties(name, SimplyStored::Couch::HasMany::Property)
         properties << SimplyStored::Couch::HasMany::Property.new(self, name, options)
       end
 
@@ -21,11 +22,12 @@ module SimplyStored
             descending = false
           end
 
-          cached_results = cached_results = send("_get_cached_#{name}")
+          cached_results = send("_get_cached_#{name}")
           cache_key = _cache_key_for(local_options)
           if forced_reload || cached_results[cache_key].nil? 
             cached_results[cache_key] = find_associated(options[:class_name], self.class, :with_deleted => with_deleted, :limit => limit, :descending => descending, :foreign_key => options[:foreign_key])
             instance_variable_set("@#{name}", cached_results)
+            self.class.set_parent_has_many_association_object(self, cached_results[cache_key])
           end
           cached_results[cache_key]
         end
@@ -149,6 +151,14 @@ module SimplyStored
         end
       end
 
+      def set_parent_has_many_association_object(parent, child_collection)
+        child_collection.each do |child|
+          if child.respond_to?("#{parent.class.name.to_s.singularize.downcase}=")
+            child.send("#{parent.class.name.to_s.singularize.camelize.downcase}=", parent)
+          end
+        end
+      end
+      
       def class_from_property_name(name)
         klass = nil
         if (properties.list && name)
@@ -171,7 +181,7 @@ module SimplyStored
             :dependent => :nullify,
             :through => nil,
             :class_name => name.to_s.singularize.camelize,
-            :foreign_key => owner_clazz.name.singularize.underscore.foreign_key
+            :foreign_key => nil
           }.update(options)
           @name, @options = name, options